requirements.txt
discord.py==1.3.4
git+https://github.com/kkroening/ffmpeg-python
PyNaCl==1.3.0
pandas
dnspython==1.16.0
async-timeout==3.0.1
requests
youtube_dl
opuslib
validators
django